Tighten url regex
The existing url regex is super permissable as to what is considered a
domain slug (e.g. any non-whitespace characters were considered
domains).

This work updates the regex to be compliant with the ietf domain name
spec, and removes erroneous test cases that were testing for invalid
regexes.

Note that this will also mean that _Invalid urls with valid punycode
encodings will no longer work_. Is this an acceptable tradeoff?