Loosen callback type requirements in chainSanitizerCallback, update dts.js for arrows in type args (#595)
Most of the time in javascript, a caller is allowed to pass a callback that is looser than the destination type. For example, you are not always required to declare the index variable someArray.map((x: T, i: number) => ...) when mapping over an array.

However, when chainSanitizerCallback is called with a record of functions of form (a:A,b:B,c:C) => D, and a callback to chain of form (a:A, b:B) => D, it throws a type error. This is because the way the function signature is currently declared enforces the argument list be exactly the same as the members of the callback record it is passed.

This change allows for callbacks that are a subtype of the record's function type, which brings behaviour in line with the the javascript ecosystem.

This change also fixes a bug in dts.js where it would fail to match functions with => in their type arguments

Co-authored-by: Maxwell Huang-Hobbs <mahuangh@microsoft.com>