Add files in clipboard (#852)
# Intent
Users can copy a file and paste it in editor which the hosts can decide how to handle it. Currently the `ClipboardData` and corresponding extraction of items does not handle this scenario.

# Change
The change is to add a new `files` field in `ClipboardData` and when extracting data we get the list of files in the clipboard and pass that information so hosts can decide how to handle it. As user can pick multiple files hence the `files` field is an `array`.

## Future Enhancements
Currently the `image` in `ClipboardData` only takes one image and ignores the next which means if user copies multiple images from computer and paste we don't add them as inline image.
With this change at least the next images will be added as files so host has an opportunity to handle them. Future improvement can involve making `image` field be an `images: File[]` so that it can handle multiple images, which should allow to insert multiple images inline in editor. 

As that would be a breaking change and requires multiple changes to existing behavior hence that is out of scope for this PR.