Add data to the ContentChange event when pasting (#1862)
When using paste we need to return the clipboardData when invoking addUndoSnapshot. That way when other plugins handle the ContentChangeEvent. They can also use the data property.
This is used for example in the PasteOption Plugin.

This PR adds this functionality to the new paste public API

Minor fix in promoteContentModelEditorCore since we were always adding the ContentModelCopyPastePlugin