Add image hidden marker (#3021)
Instead of using a dataset to store the isEditing property, a hidden property is now used. To support this, get/set functions and the ImageMarkerFormat were introduced. The imageMarker property can now be accessed through the format property of the image.
This change eliminates the need to manually remove the dataset from the image element when extracting content from the DOM.